Refaktor. Fjern overfløig css og strukturer innhold i komponenter
