+++ app/komponenter/bildeScanningGuide/bildeScanningEksempel.module.css
@@ -1,17 +1,26 @@
-.bildeKonteiner {
-  padding-left: 1rem;
-  margin-bottom: 0.75rem;
+.konteiner {
+  flex: 1 1 50%;
+  height: auto;
+  padding: 0.5rem 0.5rem 1.5rem 0.5rem;
+
+  @media screen and (min-width: 768px) {
+    max-width: 25%;
+  }
+}
+
+.konteiner:first-child {
+  padding-left: 0;
+}
+
+.konteiner:last-child {
+  padding-right: 0;
 }
 
 .statusTekst {
   display: flex;
   align-items: center;
+  gap: 0.5rem;
   margin-bottom: 0.5rem;
-}
-
-.statusIkonKonteiner {
-  padding-right: 0.5rem;
-  line-height: 0;
 }
 
 .beskrivelse {
