+++ app/komponenter/bildeScanningGuide/bildeScanningGuide.module.css
@@ -3,21 +3,3 @@
   flex-flow: row wrap;
   justify-content: space-around;
 }
-
-.bildeEksempelKonteiner {
-  flex: 1 1 50%;
-  height: auto;
-  padding: 0.5rem 0.5rem 1.5rem 0.5rem;
-
-  @media screen and (min-width: 768px) {
-    max-width: 25%;
-  }
-}
-
-.bildeEksempelKonteiner:first-child {
-  padding-left: 0;
-}
-
-.bildeEksempelKonteiner:last-child {
-  padding-right: 0;
-}
