+++ app/sider/Dokumentasjon.tsx
@@ -2,7 +2,7 @@ import { Button } from '@navikt/ds-react';
 import { useNavigate } from '@remix-run/react';
 
 import { useTekster, useYtelse } from '~/hooks/contextHooks';
-import BildeScanningGuide from '~/komponenter/bildeScanningGuide/BildeScanningGuide';
+import TaGodtBildeInfo from '~/komponenter/bildeScanningGuide/TaGodtBilde';
 import HovedInnhold from '~/komponenter/hovedInnhold/HovedInnhold';
 import StegIndikator from '~/komponenter/stegindikator/StegIndikator';
 import TekstBlokk from '~/komponenter/tekstblokk/TekstBlokk';
@@ -30,7 +30,7 @@ export default function DokumentasjonSide() {
         tekstblokk={tekster.dokumentasjonOverskrift}
         typografi={ETypografiTyper.STEG_HEADING_SMALL_H1}
       />
-      <BildeScanningGuide />
+      <TaGodtBildeInfo />
       <div className={`${css.navigeringsKnappKonteiner}`}>
         <Button
           type="button"
